refactor(agent): refold execution files split for the 250-LOC cap
Reverses the artificial 17-file split of core/execution by cohesion.
Four fragments fold into their only src consumer (tool-policy-dispatch
-> tool-executor, prompt-policy -> turn-prepare, completion-policy ->
turn-outcome, policy-engine-builder -> runner); three shared leaf
fragments merge into the sibling module their consumers already import
(run-result -> run-events, policy-effects-apply -> policy-effects,
lifecycle-context -> run-state). Symbols that became file-internal are
unexported (dispatchToolPre/Post, buildTurnSystemPrompt,
dispatchPostRunTransform, applyPostCompaction) so the dead-export
ratchet stays at 32. Two test files renamed to match their imports.
17 -> 10 files, no behavior change: 3507 pass / 0 fail, identical
test count; fold verified body-identical against HEAD^ by adversarial
review.
