+++ packages/agent/src/core/execution/lifecycle-dispatch.ts
@@ -1,13 +1,20 @@
 import { PolicyDecision, type Run } from "@openomni/protocol";
-import { effectOf } from "./policy-effects";
+import { effectOf, PolicyEffectApplier } from "./policy-effects";
 import { checkBudget } from "../budget";
 import type { PolicyEngineInstance } from "../policy";
 import type { AgentEvent, ChatAgentConfig } from "../types";
-import { emitRunCompleted, publishDenyDiagnostic } from "./run-events";
-import { PolicyEffectApplier } from "./policy-effects-apply";
-import { agentBaseForState, buildLifecyclePolicyContext } from "./lifecycle-context";
-import { createGuardCompleteEvent, createRunCompleteEvent } from "./run-result";
-import type { AgentRunBase, RunState } from "./run-state";
+import {
+  createGuardCompleteEvent,
+  createRunCompleteEvent,
+  emitRunCompleted,
+  publishDenyDiagnostic,
+} from "./run-events";
+import {
+  agentBaseForState,
+  buildLifecyclePolicyContext,
+  type AgentRunBase,
+  type RunState,
+} from "./run-state";
 
 export async function dispatchPreRun(
   state: RunState,
