+++ packages/agent/test/core/execution/policy-effects.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from "bun:test";
 import { PolicyDecision } from "@openomni/protocol";
-import { PolicyEffectApplier } from "../../../src/core/execution/policy-effects-apply";
+import { PolicyEffectApplier } from "../../../src/core/execution/policy-effects";
 import { createRunState } from "../../../src/core/execution/run-state";
 
 describe("PolicyEffectApplier", () => {
