+++ packages/agent/test/core/execution/runner-policy-engine.test.ts
@@ -1,7 +1,7 @@
 import { describe, expect, it, mock } from "bun:test";
 import { Bus } from "@openomni/session";
 import { abortRun } from "../../helpers/policy-decision";
-import { buildPolicyEngine } from "../../../src/core/execution/policy-engine-builder";
+import { buildPolicyEngine } from "../../../src/core/execution/runner";
 import { makeAgentBase, makeConfig } from "./lifecycle-dispatch-fixture";
 
 const validToolNativePreContext = {
